Move the remaining bulk of `ImmuTable` in package
It was easier than identifying the access control relationships, change
them for the purpose of moving the files, then updating once the move
was completed.
