+++ Modules/Sources/ImmuTable/Hashable+Xcode16Workaround.swift
@@ -1,0 +1,3 @@
+// This conformance was added during the Xcode 16 migration to silence the
+// dozens of false-positive warnings (any @unchecked conformance is tech debt).
+extension AnyHashable: @retroactive @unchecked Sendable {}
