+++ Modules/Sources/ImmuTable/ImmuTable+Empty.swift
@@ -1,0 +1,6 @@
+public extension ImmuTable {
+    /// Alias for an ImmuTable with no sections
+    static var Empty: ImmuTable {
+        return ImmuTable(sections: [])
+    }
+}
