+++ Modules/Sources/ImmuTable/TableViewContainer.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+@objc public protocol TableViewContainer: AnyObject {
+    var tableView: UITableView! { get }
+}
