+++ Modules/Sources/ImmuTable/UITableViewController+TableViewContainer.swift
@@ -1,0 +1,3 @@
+import UIKit
+
+extension UITableViewController: TableViewContainer {}
