+++ Modules/Sources/ImmuTable/UIViewControllerWithTableView.swift
@@ -1,0 +1,3 @@
+import UIKit
+
+public typealias UIViewControllerWithTableView = TableViewContainer & UITableViewDataSource & UITableViewDelegate & UIViewController
