+++ WordPress/Classes/ViewRelated/Stats/Shared Views/Stats Detail/SiteStatsInsightsDetailsViewModel.swift
@@ -1,4 +1,5 @@
 import Foundation
+import ImmuTable
 import UIKit
 import WordPressFlux
