+++ WordPress/Classes/ViewRelated/Stats/Subscribers/StatsSubscribersViewModel.swift
@@ -1,5 +1,6 @@
-import Foundation
 import Combine
+import Foundation
+import ImmuTable
 import WordPressKit
 
 final class StatsSubscribersViewModel {
